+++ packages/cli/src/options.ts
@@ -4,6 +4,8 @@ export type SharedOptions = {
   readonly file?: string;
   readonly tokenUsage?: boolean;
   readonly headless?: boolean;
+  readonly userSession?: boolean;
+  readonly pageFilter?: 'simple' | 'none' | 'readability';
 };
 
 export const sharedOptions = {
@@ -34,5 +36,17 @@ export const sharedOptions = {
     type: 'boolean',
     description: 'Use browser in headless mode with no UI showing',
     default: true,
+  } as const,
+  userSession: {
+    type: 'boolean',
+    description:
+      "Use user's existing browser session instead of sandboxed session",
+    default: false,
+  } as const,
+  pageFilter: {
+    type: 'string',
+    description: 'Method to process webpage content',
+    default: 'none',
+    choices: ['simple', 'none', 'readability'],
   } as const,
 };